P04Search/requirements.txt
Django~=5.0.4
jieba~=0.42.1
django-import-export~=4.0.0rc3
pandas~=2.2.2
tqdm~=4.66.2
openpyxl~=3.1.2
Levenshtein~=0.25.1